refactor(error): make error trace more generic
Move everything to a submodule to horizontalize development.
