+++ src/error/trace.rs
@@ -1,9 +1,10 @@
+use super::{pest_error::PestError, span::Span};
+
 use crate::parsing::*;
 
 use pest::{
     error::{Error, ErrorVariant, LineColLocation},
     iterators::{Pair, Pairs},
-    Span,
 };
 
 use std::fmt::Debug;
@@ -17,11 +18,6 @@ pub enum Stage {
     Compiling,
 }
 
-#[derive(Debug, Default)]
-pub struct Trace {
-    stack: Vec<(Stage, Box<dyn TraceError>)>,
-}
-
 pub trait TraceError: Debug {
     fn from_span(span: Span, message: &str) -> Self
     where
@@ -29,6 +25,11 @@ pub trait TraceError: Debug {
     fn line_col(&self) -> LineColLocation;
     fn line(&self) -> &str;
     fn message(&self) -> &str;
+}
+
+#[derive(Debug, Default)]
+pub struct Trace {
+    stack: Vec<(Stage, Box<dyn TraceError>)>,
 }
 
 impl<T: TraceError + 'static> From<(Stage, T)> for Trace {
@@ -42,7 +43,7 @@ impl<T: TraceError + 'static> From<(Stage, T)> for Trace {
 impl<T: TraceError + 'static> From<T> for Trace {
     fn from(err: T) -> Self {
         Trace {
-            stack: vec![(Stage::Parsing, Box::new(err))],
+            stack: vec![(Stage::Unknown, Box::new(err))],
         }
     }
 }
@@ -54,9 +55,9 @@ impl Trace {
         }
     }
 
-    pub fn new_from_message(stage: Stage, pair: &Pair<Rule>, message: String) -> Self {
+    pub fn new_from_pair(pair: &Pair<Rule>, message: String) -> Self {
         let mut res = Trace::default();
-        res.push_message(stage, pair, message);
+        res.push_pest_error(Stage::Parsing, pair, message);
         res
     }
 
@@ -64,53 +65,14 @@ impl Trace {
         self.stack.push((stage, Box::new(err)))
     }
 
-    pub fn push_message(&mut self, stage: Stage, pair: &Pair<Rule>, message: String) {
+    pub fn push_pest_error(&mut self, stage: Stage, pair: &Pair<Rule>, message: String) {
         self.stack.push((
             stage,
-            Box::new(PestError::from_span(pair.as_span(), message.as_ref())),
+            Box::new(PestError::from_span(
+                pair.as_span().into(),
+                message.as_ref(),
+            )),
         ))
-    }
-}
-
-#[derive(Debug)]
-pub struct PestError {
-    line_col: LineColLocation,
-    line: String,
-    message: String,
-}
-
-impl TraceError for PestError {
-    fn from_span(span: Span, message: &str) -> Self
-    where
-        Self: Sized,
-    {
-        Self {
-            line_col: LineColLocation::Span(span.start_pos().line_col(), span.end_pos().line_col()),
-            line: span.as_str().to_owned(),
-            message: message.to_owned(),
-        }
-    }
-
-    fn line_col(&self) -> LineColLocation {
-        self.line_col.clone()
-    }
-
-    fn line(&self) -> &str {
-        self.line.as_ref()
-    }
-
-    fn message(&self) -> &str {
-        self.message.as_ref()
-    }
-}
-
-impl From<Error<Rule>> for PestError {
-    fn from(err: Error<Rule>) -> Self {
-        Self {
-            line_col: err.line_col.clone(),
-            line: err.line().to_owned(),
-            message: err.variant.message().to_string(),
-        }
     }
 }
 
@@ -133,7 +95,6 @@ impl std::fmt::Display for Trace {
 
                     let arrow = format!("{}>", "-".repeat(line_nbr_len));
 
-                    // Error coordinates
                     let coords = match err.line_col() {
                         LineColLocation::Pos((y, x)) => format!("{y}:{x}"),
                         LineColLocation::Span((ys, xs), (ye, xe)) => {
@@ -141,7 +102,6 @@ impl std::fmt::Display for Trace {
                         }
                     };
 
-                    // Underline
                     let underline = match err.line_col() {
                         LineColLocation::Pos((_, x)) => format!("{}^", " ".repeat(x)),
                         LineColLocation::Span((ys, xs), (ye, xe)) => {
